requirements.txt
fal-client>=0.4.0
pillow>=12.1.0
rembg[cpu]>=2.0.0
opencv-python-headless>=4.8.0
numpy>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
